+++ src/intergalactic/tests/test_dtds.py
@@ -2,6 +2,7 @@ import pytest
 import numpy as np
 import intergalactic.settings as settings
 from intergalactic.dtds import select_dtd
+from intergalactic.dtds import dtd_correction
 from intergalactic.dtds import dtd_ruiz_lapuente
 from intergalactic.dtds import dtd_maoz_graur
 from intergalactic.dtds import dtd_castrillo
@@ -25,7 +26,7 @@ def test_select_dtd(available_dtds):
     dtds = [dtd_ruiz_lapuente, dtd_maoz_graur, dtd_castrillo, dtd_greggio]
 
     for i in range(len(available_dtds)):
-        times = [0.001, 9.] + list(np.random.rand(5)) + list(np.random.rand(5) * 9)
+        times = [0.001, 0.04, 0.1, 0.4, 9.] + list(np.random.rand(5)) + list(np.random.rand(5) * 9)
         for time in times:
             assert select_dtd(available_dtds[i])(time) == dtds[i](time)
 
@@ -34,3 +35,8 @@ def test_no_negative_time_values():
     t = -1
     assert dtd_ruiz_lapuente(t) == 0.0
     assert dtd_maoz_graur(t) == 0.0
+
+
+def test_dtd_correction_factor():
+    assert dtd_correction({}) == 1.0
+    assert dtd_correction({'dtd_correction_factor': 3.0}) == 3.0